feat(images): add default images variations
Add a new method  to set default values for . This method is used in the  and  functions to set default values for , , , and  fields of the  struct.

In the  function, updated the code to use the default values returned by the  method.

In the  function, also updated the code to use the default values returned by the  method, and added error handling when making the API request.

Additionally, removed the unnecessary assignment of  in the  function.

Resolves: Issue #123