feat(images): add error handling for HTTP status code
Add error handling for HTTP status code in the ImagesGenerations, ImagesGenerationsB64JSON, ImagesVariations, and ImagesVariationsB64JSON functions. If the HTTP status code is not 200 OK, the functions will return an ErrorResponse object with the error message.

In the ImagesGenerations function, if the request fails, the function will return the ErrorResponse object instead of just nil. Similarly, in the ImagesGenerationsB64JSON, ImagesVariations, and ImagesVariationsB64JSON functions, the ErrorResponse object is returned in case of failure.

Additionally, when making the HTTP request, the functions now set the Error field in the ErrorResponse object using the SetError method.

Lastly, after making the HTTP request, if the status code is not 200 OK, the functions return the ErrorResponse object instead of nil.

This change improves the error handling in the images package and provides more informative error messages when encountering non-200 HTTP